class EdfDataProvider are ready to be tested

+++ edf-viewer/src/main/java/com/biorecorder/edfviewer/DataListener.java
@@ -1,7 +1,7 @@
 package com.biorecorder.edfviewer;
 
 public interface DataListener {
-    public void onDataReceived(int data);
+    public void onDataReceived(double data);
     public void onFinish();
     public void onStart();
    // public int getFilterLength();
